java utility and JDK dir. Harness actually tries to launch it.
