+++ src/IKVM.JavaTest.AgentServer/IKVM.JavaTest.AgentServer.csproj
@@ -2,6 +2,7 @@
     <Import Sdk="Microsoft.NET.Sdk" Project="Sdk.props"/>
     <Import Project="$(MSBuildThisFileDirectory)..\IKVM-pkg\buildTransitive\netstandard2.0\IKVM.props" />
     <Import Project="$(MSBuildThisFileDirectory)..\IKVM.MSBuild\buildTransitive\netstandard2.0\IKVM.MSBuild.props" />
+    <Import Project="$(MSBuildThisFileDirectory)..\..\jtreg.props" />
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
@@ -45,9 +46,8 @@
     </ItemGroup>
 
     <ItemGroup>
-        <IkvmReference Include="..\..\jtreg\build\deps\**\*.jar" />
+        <IkvmReference Include="@(JTRegJar)" />
     </ItemGroup>
-
 
     <Import Sdk="Microsoft.NET.Sdk" Project="Sdk.targets"/>
     <Import Project="$(MSBuildThisFileDirectory)..\IKVM-pkg\buildTransitive\netstandard2.0\IKVM.targets" />
