+++ src/IKVM.JavaTest.TestAdapter.Tests/IKVM.JavaTest.TestAdapter.Tests.csproj
@@ -2,6 +2,7 @@
     <Import Sdk="Microsoft.NET.Sdk" Project="Sdk.props"/>
     <Import Project="$(MSBuildThisFileDirectory)..\IKVM-pkg\buildTransitive\netstandard2.0\IKVM.props" />
     <Import Project="$(MSBuildThisFileDirectory)..\IKVM.MSBuild\buildTransitive\netstandard2.0\IKVM.MSBuild.props" />
+    <Import Project="$(MSBuildThisFileDirectory)..\..\jtreg.props" />
 
     <PropertyGroup>
         <TargetFrameworks>net461;netcoreapp3.1</TargetFrameworks>
@@ -51,7 +52,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <IkvmReference Include="..\..\jtreg\build\deps\**\*.jar" />
+        <IkvmReference Include="@(JTRegJar)" />
     </ItemGroup>
 
     <ItemGroup>
