+++ src/IKVM.JavaTest.TestAdapter/IKVM.JavaTest.TestAdapter.csproj
@@ -2,6 +2,7 @@
     <Import Sdk="Microsoft.NET.Sdk" Project="Sdk.props"/>
     <Import Project="$(MSBuildThisFileDirectory)..\IKVM-pkg\buildTransitive\netstandard2.0\IKVM.props" />
     <Import Project="$(MSBuildThisFileDirectory)..\IKVM.MSBuild\buildTransitive\netstandard2.0\IKVM.MSBuild.props" />
+    <Import Project="$(MSBuildThisFileDirectory)..\..\jtreg.props" />
 
     <PropertyGroup>
         <TargetFrameworks>net461;netcoreapp3.1</TargetFrameworks>
@@ -71,8 +72,34 @@
         </PublishProjectReference>
     </ItemGroup>
 
+    <ItemGroup Condition=" '$(TargetFramework)' == 'net461' ">
+        <PublishProjectReference Include="..\java\java.csproj">
+            <SetTargetFramework>TargetFramework=net461</SetTargetFramework>
+            <PublishTargetPath>jdk\bin</PublishTargetPath>
+            <PublishProperties></PublishProperties>
+        </PublishProjectReference>
+        <PublishProjectReference Include="..\java\java.csproj">
+            <SetTargetFramework>TargetFramework=net461</SetTargetFramework>
+            <PublishTargetPath>jdk\bin</PublishTargetPath>
+            <PublishProperties></PublishProperties>
+        </PublishProjectReference>
+    </ItemGroup>
+
+    <ItemGroup Condition=" '$(TargetFramework)' == 'netcoreapp3.1' ">
+        <PublishProjectReference Include="..\java\java.csproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <PublishTargetPath>jdk\bin</PublishTargetPath>
+            <PublishProperties></PublishProperties>
+        </PublishProjectReference>
+        <PublishProjectReference Include="..\javac\javac.csproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <PublishTargetPath>jdk\bin</PublishTargetPath>
+            <PublishProperties></PublishProperties>
+        </PublishProjectReference>
+    </ItemGroup>
+
     <ItemGroup>
-        <IkvmReference Include="..\..\jtreg\build\deps\**\*.jar" />
+        <IkvmReference Include="@(JTRegJar)" />
     </ItemGroup>
 
     <Import Sdk="Microsoft.NET.Sdk" Project="Sdk.targets"/>
