+++ src/IKVM.Tool.Launcher/IKVM.Tool.Launcher.csproj
@@ -2,7 +2,6 @@
 
     <PropertyGroup>
         <TargetFrameworks>net461;netcoreapp3.1</TargetFrameworks>
-        <DefineConstants>$(DefineConstants);STUB_GENERATOR</DefineConstants>
     </PropertyGroup>
 
     <ItemGroup>
@@ -38,35 +37,9 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\IKVM.Reflection\IKVM.Reflection.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <Compile Include="..\IKVM.Runtime\Extensions\**\*.cs" LinkBase="runtime\Extensions" />
-        <Compile Include="..\IKVM.Runtime\Syntax\**\*.cs" LinkBase="runtime\Syntax" />
-        <Compile Include="..\IKVM.Runtime\InternalException.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\AssemblyClassLoader.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\Attributes\**\*.cs" LinkBase="runtime\Attributes" />
-        <Compile Include="..\IKVM.Runtime\ClassLoaderWrapper.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\GenericClassLoaderWrapper.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\CoreClasses.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\DotNetTypeWrapper.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\JavaException.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\MemberWrapper.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\profiler.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\ReflectUtil.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\tracer.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\Types.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\TypeWrapper.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\TypeWrapperFactory.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\vm.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\StubGen\**\*.cs" LinkBase="runtime\SubGen" />
-        <Compile Include="..\IKVM.Runtime\LoadMode.cs" LinkBase="runtime" />
-        <Compile Include="..\IKVM.Runtime\CodeGenOptions.cs" LinkBase="runtime" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <Compile Include="..\ikvmc\IKVM\Internal\AssemblyResolver.cs" LinkBase="ikvmc\Internal" />
+        <ProjectReference Include="..\IKVM.Java\IKVM.Java.msbuildproj" />
+        <ProjectReference Include="..\IKVM.Runtime\IKVM.Runtime.csproj" />
+        <ProjectReference Include="..\IKVM.Runtime.JNI\IKVM.Runtime.JNI.csproj" />
     </ItemGroup>
 
 </Project>
