+++ src/ikvm/Program.cs
@@ -1,0 +1,14 @@
+namespace ikvm
+{
+
+    public static class Program
+    {
+
+        public static int Main(string[] args)
+        {
+            return java.Program.Main(args);
+        }
+
+    }
+
+}
