+++ src/java/java.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFrameworks>net461;netcoreapp3.1</TargetFrameworks>
+        <RuntimeIdentifiers>win7-x64;linux-x64</RuntimeIdentifiers>
+        <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\IKVM.Tool.Launcher\IKVM.Tool.Launcher.csproj" />
+    </ItemGroup>
+
+</Project>
